fix(npmtag): deprecate usage of npmtag for artifacts
Fixes #687, npmTag option will be deprecated across sfpowerscripts to reduce complexity while
handling artifacts. All artifacts will be picked up using LATEST_GIT_TAG that can be reached from a
given branch

BREAKING CHANGE: Pool Schema no longer supports npmTag

fix #687
